pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-podcast-producer"
version = "1.0.0"
description = "Automated workflow to convert NotebookLM audio into fully produced video podcasts"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Israel Castro", email = "soyisracastro@gmail.com"}
]
keywords = [
    "podcast",
    "ai",
    "automation",
    "video-production",
    "speaker-diarization",
    "transcription",
    "notebooklm",
    "heygen",
    "content-creation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Content Creators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    # AI & Diarization
    "pyannote.audio",
    "torch",
    "torchaudio",

    # Transcription
    "openai-whisper",

    # Content Analysis
    "openai",

    # Audio & Video Processing
    "pydub",
    "moviepy",
    "imageio-ffmpeg",

    # Configuration
    "python-dotenv",

    # Notion Integration
    "notion-client",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.12",
    "mypy>=1.0",
    "ruff>=0.1",
]
cli = [
    "typer>=0.9",
    "rich>=13.0",
]
web = [
    "streamlit>=1.28",
]
all = [
    "ai-podcast-producer[dev,cli,web]",
]

[project.urls]
Homepage = "https://github.com/soyisracastro/ai-podcast-producer"
Documentation = "https://github.com/soyisracastro/ai-podcast-producer#readme"
Repository = "https://github.com/soyisracastro/ai-podcast-producer.git"
Issues = "https://github.com/soyisracastro/ai-podcast-producer/issues"
Changelog = "https://github.com/soyisracastro/ai-podcast-producer/blob/master/CHANGELOG.md"

[project.scripts]
# Future CLI commands (when cli extra is installed)
# podcast-producer = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt", "*.json"]

# === Development Tools Configuration ===

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | input
  | output
  | archives
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = ["venv", ".venv", "input", "output", "archives"]

[tool.ruff]
line-length = 100
target-version = "py39"
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    "venv",
    ".venv",
    "input",
    "output",
    "archives",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    "venv",
    "input",
    "output",
    "archives",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
